✨ feat: 호스트 \& 참여자 각 화면 컨테이너화
